TST: Move tests folder to root and enable running them all at once
