+++ Dockerfile
@@ -37,6 +37,6 @@ COPY . /workspaces/workoutizer
 WORKDIR /workspaces/workoutizer
 
 # set convenience alias
-RUN echo 'alias run_all_tests="pytest wkz/tests -v -n4 --html=pytest-report.html"' >> ~/.zshrc
+RUN echo 'alias run_all_tests="pytest tests -v -n4 --html=pytest-report.html"' >> ~/.zshrc
 
 RUN /bin/bash -c 'source /tmp/venv/bin/activate && pip install -e . --no-deps --disable-pip-version-check'
