+++ tests/data/swim_no_coordinates.fit
