+++ tests/data/template_example.gpx
