+++ tests/db_tests/browser_tests/test_activity_page.py
