+++ tests/db_tests/browser_tests/test_help_page.py
