+++ tests/db_tests/browser_tests/test_settings_page.py
@@ -3,13 +3,11 @@ import operator
 
 from django.urls import reverse
 from selenium.common.exceptions import NoSuchElementException
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.common.by import By
 import pytest
 
 from workoutizer import settings as django_settings
-from wkz.tests.utils import delayed_assertion
+from tests.utils import delayed_assertion
 from wkz import models
 
 
@@ -90,7 +88,6 @@ def test_settings_page__demo_activity_present__delete_it(import_demo_data, live_
         webdriver.find_element(By.ID, "delete-demo-data")
 
 
-@pytest.mark.no_autouse
 def test_settings_page__edit_and_submit_form(live_server, webdriver):
     # get settings and check that all values are at their default configuration
     settings = models.get_settings()
@@ -140,32 +137,3 @@ def test_settings_page__edit_and_submit_form(live_server, webdriver):
     # this attribute got removed, so verifying that is does no longer exist
     with pytest.raises(AttributeError):
         assert settings.reimporter_updates_all is True
-
-
-def test_settings_page__reimport_activities(live_server, webdriver, import_one_activity):
-    import_one_activity("cycling_bad_schandau.fit")
-    assert models.Activity.objects.count() == 1
-    activity = models.Activity.objects.get()
-    original_distance = activity.distance
-
-    # change some values
-    activity.name = "Foo"
-    activity.distance = 999.9
-    activity.save()
-
-    activity = models.Activity.objects.get()
-    assert activity.name == "Foo"
-
-    webdriver.get(live_server.url + reverse("settings"))
-    assert webdriver.find_element_by_class_name("navbar-brand").text == "Settings"
-
-    # trigger reimport
-    webdriver.find_element(By.ID, "reimport-activities").click()
-
-    # check that loading bar image is present
-    WebDriverWait(webdriver, 3).until(EC.presence_of_element_located((By.ID, "loading-bar")))
-
-    # the name should not have changed during reimporting
-    delayed_assertion(models.Activity.objects.filter(name="Foo").count, operator.eq, 1)
-    # the distance however should have been reverted to the original value
-    delayed_assertion(models.Activity.objects.filter(distance=original_distance).count, operator.eq, 1)
