+++ tests/db_tests/conftest.py
@@ -1,5 +1,9 @@
-import datetime
 import os
+import datetime
+import shutil
+from py._path.local import LocalPath
+from pathlib import Path
+from typing import List
 
 import pytest
 import pytz
@@ -12,6 +16,16 @@ from wkz.file_importer import (
     copy_demo_fit_files_to_track_dir,
 )
 from wkz import models
+
+
+@pytest.fixture
+def tracks_in_tmpdir(db, tmp_path):
+    path = tmp_path / "test_traces"
+    path.mkdir()
+    settings = models.get_settings()
+    settings.path_to_trace_dir = str(path)
+    settings.save()
+    return settings
 
 
 @pytest.fixture
@@ -67,17 +81,6 @@ def activity(db, sport, trace_file):
     return activity
 
 
-@pytest.fixture
-def test_data_dir():
-    return os.path.join(os.path.dirname(__file__), "data")
-
-
-@pytest.fixture
-def demo_data_dir():
-    return django_settings.INITIAL_TRACE_DATA_DIR
-
-
-@pytest.fixture(autouse=True)
 def ensure_path_to_traces_is_tmp(db, monkeypatch, tmp_path, request):
     def get_dummy_settings(*args, **kwargs):
         target_dir = tmp_path / "traces"
@@ -99,7 +102,7 @@ def ensure_path_to_traces_is_tmp(db, monkeypatch, tmp_path, request):
 
 
 @pytest.fixture
-def import_demo_data(db):
+def import_demo_data(tracks_in_tmpdir):
     prepare_import_of_demo_activities(models)
     assert models.Sport.objects.count() == 5
     assert models.Settings.objects.count() == 1
@@ -109,7 +112,7 @@ def import_demo_data(db):
 
 
 @pytest.fixture
-def import_one_activity(db, test_data_dir, demo_data_dir):
+def import_one_activity(tracks_in_tmpdir, test_data_dir, demo_data_dir):
     models.get_settings()
     assert models.Settings.objects.count() == 1
 
@@ -184,3 +187,94 @@ def insert_best_section(db, activity, insert_activity):
         return best_section
 
     return _create_section
+
+
+class FakeDevice:
+    def __init__(self, mount_path: LocalPath, device_dir, activity_dir: str, activity_files):
+        self.mount_path = mount_path
+        self.device_path = self.mount_path / device_dir
+        self.activity_path_on_device = self.device_path / activity_dir
+        self.mounted = False
+        self.activity_files = activity_files
+
+    def mount(self):
+        # create directories for activities
+        if not self.mounted:
+            print("mounting fake device...")
+            self.activity_path_on_device.mkdir(parents=True)
+            if self.activity_files:
+                # copy activity files into activity dir
+                copy_demo_fit_files_to_track_dir(
+                    source_dir=django_settings.INITIAL_TRACE_DATA_DIR,
+                    targe_dir=self.activity_path_on_device,
+                    list_of_files_to_copy=self.activity_files,
+                )
+            self.mounted = True
+
+    def unmount(self):
+        if self.mounted:
+            print("unmounting fake device...")
+            shutil.rmtree(self.device_path)
+            self.mounted = False
+
+
+@pytest.fixture
+def fake_device(tmp_path, device_dir, activity_dir):
+    settings = models.get_settings()
+
+    mount_path = tmp_path / "mount_path"
+    mount_path.mkdir()
+    assert mount_path.is_dir()
+    settings.path_to_garmin_device = mount_path
+
+    trace_dir = tmp_path / "trace_dir"
+    trace_dir.mkdir()
+    assert trace_dir.is_dir()
+    settings.path_to_trace_dir = trace_dir
+
+    settings.save()
+
+    def _get_device(activity_files: List[str], device_dir: str = device_dir, activity_dir: str = activity_dir):
+        return FakeDevice(
+            mount_path=settings.path_to_garmin_device,
+            device_dir=device_dir,
+            activity_dir=activity_dir,
+            activity_files=activity_files,
+        )
+
+    return _get_device
+
+
+@pytest.fixture
+def fit_file():
+    return "2019-09-14-17-22-05.fit"
+
+
+@pytest.fixture
+def fit_file_a():
+    return "2019-09-18-16-02-35.fit"
+
+
+@pytest.fixture
+def fit_file_b():
+    return "2019-09-25-16-15-53.fit"
+
+
+@pytest.fixture
+def device_dir():
+    return "mtp:host_DEVICE"
+
+
+@pytest.fixture
+def activity_dir():
+    return "Primary/GARMIN/Activity"
+
+
+@pytest.fixture
+def import_sequentially_on_setting_save(monkeypatch):
+    def import_sequentially(self, *args, **kwargs):
+        super(models.Settings, self).save(*args, **kwargs)
+        if Path(self.path_to_trace_dir).is_dir():
+            run_file_importer(models, as_huey_task=False)
+
+    monkeypatch.setattr(models.Settings, "save", import_sequentially)
