+++ tests/db_tests/test_automatic_naming.py
