+++ tests/db_tests/test_cli.py
@@ -18,7 +18,7 @@ def test_cli_version():
     assert output == f"{__version__}\n"
 
 
-def test_cli__init(db):
+def test_cli__init(tracks_in_tmpdir):
     cli._init(import_demo_activities=False)
     assert os.path.isdir(django_settings.WORKOUTIZER_DIR)
     assert len(models.Sport.objects.all()) == 0
@@ -32,7 +32,7 @@ def test_cli__init(db):
     assert len(models.Activity.objects.all()) > 1
 
 
-def test_cli__check__demo_data(db):
+def test_cli__check__demo_data(tracks_in_tmpdir):
     # initialized wkz with demo data and then run check
     cli._init(import_demo_activities=True)
     cli._check()
@@ -51,7 +51,7 @@ def test_cli__check__not_initialized(db):
         cli._check()
 
 
-def test_cli__reimport(db, import_one_activity):
+def test_cli__reimport(import_one_activity):
     import_one_activity("cycling_bad_schandau.fit")
 
     assert models.Activity.objects.count() == 1
