+++ tests/db_tests/test_export_gpx_file.py
