+++ tests/unit_tests/file_helper/test_fit_collector.py
